+++ packages/api/src/controllers/admin/manage/Citizens.ts
@@ -4,7 +4,7 @@ import { UseBeforeEach } from "@tsed/platform-middlewares";
 import { BodyParams, Context, PathParams } from "@tsed/platform-params";
 import { Delete, Description, Get, Put } from "@tsed/schema";
 import { userProperties } from "lib/auth/user";
-import { leoProperties } from "lib/officer";
+import { leoProperties } from "lib/leo/activeOfficer";
 import { prisma } from "lib/prisma";
 import { IsAuth } from "middlewares/index";
 import { CREATE_CITIZEN_SCHEMA } from "@snailycad/schemas";
