+++ packages/api/src/controllers/admin/manage/Units.ts
@@ -10,7 +10,7 @@ import {
   unlinkDivisionsFromOfficer,
   validateMaxDivisionsPerOfficer,
 } from "controllers/leo/LeoController";
-import { leoProperties, unitProperties } from "lib/officer";
+import { leoProperties, unitProperties } from "lib/leo/activeOfficer";
 import { prisma } from "lib/prisma";
 import { validateSchema } from "lib/validateSchema";
 import { IsAuth } from "middlewares/index";
