+++ packages/api/src/controllers/dispatch/StatusController.ts
@@ -15,7 +15,7 @@ import { BodyParams, Context, PathParams } from "@tsed/platform-params";
 import { Description, Post, Put } from "@tsed/schema";
 import { prisma } from "lib/prisma";
 import { callInclude, findUnit } from "./911-calls/Calls911Controller";
-import { leoProperties, unitProperties } from "lib/officer";
+import { leoProperties, unitProperties } from "lib/leo/activeOfficer";
 import { sendDiscordWebhook } from "lib/discord/webhooks";
 import { Socket } from "services/SocketService";
 import { IsAuth } from "middlewares/index";
