+++ packages/api/src/lib/ems-fd.ts
@@ -1,7 +1,7 @@
 import type { User } from "@prisma/client";
 import type { Req, Context } from "@tsed/common";
 import { BadRequest, Forbidden, Unauthorized } from "@tsed/exceptions";
-import { unitProperties } from "./officer";
+import { unitProperties } from "lib/leo/activeOfficer";
 import { prisma } from "./prisma";
 
 export async function getActiveDeputy(req: Req, user: User, ctx: Context) {
